[gk] Тесты psql
